For now just add the ImageInformation entry when schema upgrading.
A better solution (update or insert) in the Image scanner leads to
the SQL "upsert" problem, for which I found no good solution for SQLite.

CCBUG: 226853

svn path=/trunk/extragear/graphics/digikam/; revision=1090694

+++ libs/database/schemaupdater.cpp
@@ -943,6 +943,15 @@ bool SchemaUpdater::updateV4toV5()
        )
          return false;
 
+    if (!m_access->backend()->execSql(QString(
+                    "REPLACE INTO ImageInformation "
+                    " (imageId) "
+                    "SELECT id FROM Images;"
+                                             ),
+                    DatabaseItem::Visible, DatabaseItem::UndefinedCategory)
+       )
+         return false;
+
     // remove orphan images that would not be removed by CollectionScanner
     m_access->backend()->execSql(QString("DELETE FROM Images WHERE album NOT IN (SELECT id FROM Albums);"));
